salvo2

+++ exercicio1.py
@@ -1,4 +1,4 @@
-nota = float(input("Digite a nota: "))
+nota = int(input("Digite a nota: "))
 
 if nota >= 60:
     print("aprovado")
